Rename KvSwashie to SwashieFace since it's not really a UI building block.
